organize element files a bit
