+++ include/InternalForceContribution.h
@@ -75,7 +75,7 @@ class InternalForceContribution : public ForceContribution {
     std::shared_ptr<aperi::MeshData> m_mesh_data;                     ///< The mesh data associated with the force contribution.
     std::string m_part_name;                                          ///< The name of the part associated with the force contribution.
     size_t m_num_nodes_per_element;                                   ///< The number of nodes per element.
-    std::shared_ptr<aperi::Element> m_element;                        ///< The element associated with the force contribution.
+    std::shared_ptr<aperi::ElementBase> m_element;                    ///< The element associated with the force contribution.
     std::shared_ptr<aperi::ElementProcessor<3>> m_element_processor;  ///< The element processor associated with the force contribution.
 };
